dir improvements: move `engineapi` to `execution`
