+++ cl/beacon/builder/mock_services/builder_client_mock.go
@@ -16,7 +16,7 @@ import (
 	common "github.com/erigontech/erigon-lib/common"
 	builder "github.com/erigontech/erigon/cl/beacon/builder"
 	cltypes "github.com/erigontech/erigon/cl/cltypes"
-	engine_types "github.com/erigontech/erigon/turbo/engineapi/engine_types"
+	engine_types "github.com/erigontech/erigon/execution/engineapi/engine_types"
 	gomock "go.uber.org/mock/gomock"
 )
