+++ cl/phase1/execution_client/execution_client_direct.go
@@ -32,8 +32,8 @@ import (
 	"github.com/erigontech/erigon/cl/clparams"
 	"github.com/erigontech/erigon/cl/cltypes"
 	"github.com/erigontech/erigon/cl/monitor"
+	"github.com/erigontech/erigon/execution/engineapi/engine_types"
 	"github.com/erigontech/erigon/execution/eth1/eth1_chain_reader"
-	"github.com/erigontech/erigon/turbo/engineapi/engine_types"
 )
 
 const reorgTooDeepDepth = 3
