+++ cl/phase1/execution_client/execution_engine_mock.go
@@ -19,7 +19,7 @@ import (
 	typesproto "github.com/erigontech/erigon-lib/gointerfaces/typesproto"
 	types "github.com/erigontech/erigon-lib/types"
 	cltypes "github.com/erigontech/erigon/cl/cltypes"
-	engine_types "github.com/erigontech/erigon/turbo/engineapi/engine_types"
+	engine_types "github.com/erigontech/erigon/execution/engineapi/engine_types"
 	gomock "go.uber.org/mock/gomock"
 )
