+++ cl/phase1/execution_client/types.go
@@ -16,7 +16,7 @@
 
 package execution_client
 
-import "github.com/erigontech/erigon/turbo/engineapi/engine_types"
+import "github.com/erigontech/erigon/execution/engineapi/engine_types"
 
 type PayloadStatus int
