+++ execution/eth1/forkchoice.go
@@ -38,7 +38,7 @@ import (
 	"github.com/erigontech/erigon/eth/consensuschain"
 	"github.com/erigontech/erigon/eth/stagedsync"
 	"github.com/erigontech/erigon/eth/stagedsync/stages"
-	"github.com/erigontech/erigon/turbo/engineapi/engine_helpers"
+	"github.com/erigontech/erigon/execution/engineapi/engine_helpers"
 )
 
 // This is the range in which we sanity check and potentially fix the canonical chain if it is broken.
